address PR #27 review feedback
