+++ Pala_One_2_1/src/storage/page_cache.cpp
@@ -7,7 +7,8 @@
 //
 //  File format (little-endian):
 //    uint32 magic            kPageCacheMagic
-//    uint32 layoutVersion    encodeLayoutVersion(layout)  — bodySize/lineGap/family/bionic
+//    uint32 layoutVersion    encodeLayoutVersion(layout)
+//                            — bodySize/lineGap/family/bionic/statusbarReserve
 //    uint32 fileSize         source-book size at save time
 //    uint16 count            number of entries that follow
 //    uint32 offsets[count]   byte offsets of pages 0..count-1
@@ -16,12 +17,16 @@
 //    0x50434F46 — original, no stamp
 //    0x50434F47 — added 16-bit `(bodySize, lineGap)` stamp
 //    0x50434F48 — widened stamp to 32 bits to also cover font family + bionic
+//    0x50434F49 — added statusbar-reserve byte to the 32-bit stamp; cache
+//                 must rebuild when Statusbar::setMode toggles between
+//                 Full / Minimal / Hidden because each has a different
+//                 reserve height and therefore a different maxLines
 //
 //  Old files fail the magic check, get ignored, then overwritten on the next
 //  save. No migration code needed.
 // ============================================================================
 
-static constexpr uint32_t kPageCacheMagic = 0x50434F48UL;
+static constexpr uint32_t kPageCacheMagic = 0x50434F49UL;
 
 static constexpr size_t kHeaderBytes =
     sizeof(uint32_t)   // magic
@@ -30,14 +35,16 @@ static constexpr size_t kHeaderBytes =
   + sizeof(uint16_t);  // count
 
 // Compact encoding of "what layout were the offsets in this file computed
-// under?" — bodySize ∈ {8,10,12,14}, lineGap ∈ [0,4], family ∈ {0,1},
-// bionic ∈ {0,1}. Each tucks into one byte, so a stride-by-256 packing is
-// unambiguous and trivially injective.
+// under?" — every field tucks into one byte (bodySize ∈ {8,10,12,14},
+// lineGap ∈ [0,4], family ∈ {0,1}, bionic ∈ {0,1}, statusbarReserve in
+// pixels — currently 0/1/STATUS_H). family + bionic share the third byte
+// (4 bits each) to leave room for statusbarReserve in the top byte.
 static uint32_t encodeLayoutVersion(const PageCacheLayout& layout) {
-  return ((uint32_t)(layout.bodySize & 0xFF))
-       | ((uint32_t)(layout.lineGap  & 0xFF) << 8)
-       | ((uint32_t)(layout.family   & 0xFF) << 16)
-       | ((uint32_t)(layout.bionic   & 0xFF) << 24);
+  return ((uint32_t)(layout.bodySize         & 0xFF))
+       | ((uint32_t)(layout.lineGap          & 0xFF) << 8)
+       | ((uint32_t)(layout.family           & 0x0F) << 16)
+       | ((uint32_t)(layout.bionic           & 0x0F) << 20)
+       | ((uint32_t)(layout.statusbarReserve & 0xFF) << 24);
 }
 
 static String pageCachePathForBook(const String& path) {
